More cleanup.

+++ g_misc.c
@@ -1558,8 +1558,8 @@ void SP_target_character (edict_t *self)
 void target_string_use (edict_t *self, edict_t *other, edict_t *activator)
 {
 	edict_t	*e;
-	int		n;
-	int		l;
+	size_t	n;
+	size_t	l;
 	char	c;
 
 	l = strlen(self->message);
